Add headless service for StatefulSet
In a Kubernetes StatefulSet, the serviceName field specifies the Headless Service associated with the StatefulSet. This service is responsible for network identity, including network routing and DNS naming, for the pods based on their ordinal index.

Also removed the redundant 'statefulset.name' and 'statefulset.serviceName' from README and values file as they're not used or required anymore.