Add headless service for StatefulSet (#60)
* Add headless service for StatefulSet

In a Kubernetes StatefulSet, the serviceName field specifies the Headless Service associated with the StatefulSet. This service is responsible for network identity, including network routing and DNS naming, for the pods based on their ordinal index.

Also removed the redundant 'statefulset.name' and 'statefulset.serviceName' from README and values file as they're not used or required anymore.

* Removed unnecessary type definition

* Added configuration support for the headless service

It allows overwriting the default for the `publishNotReadyAddresses` spec and adds support for adding custom annotations just for the headless service object.

* Add test for creating headless service without extra annotation

The nxrm-ha/tests/service-headless_test.yaml has been updated with a new test case. This test checks that a headless service can be created without using any additional headless annotations.

---------

Co-authored-by: Rene Schönlein <rene.schoenlein@cdi-ag.de>